Extract workspace tool-run APIs
